+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Release Notes
 
 ## Upcoming version
+
 - Adding some special votes
 - Automatic Djinn and Bootlegger
 - Updating night order
 - Updating jinxes
+- Showing pertinent Fabled during setup
 
 ### Version 3.20.1
+
 - Inconsistancies in french translations for "Power" & "Character"
 
 ### Version 3.20.0